pyproject.toml
[tool.poetry]
name = "together-bot"
version = "0.1.0"
description = ""
authors = ["Lyuha <lyuha@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
"discord.py" = "^1.5.1"
python-dotenv = "^0.14.0"
pyyaml = "^5.3.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.5.2"
mypy = "^0.782"

[tool.poetry.scripts]
bot = 'together_bot.bot:start'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
